Remove OpenTelemetryBuilder interface, every implementation will have it's own builder
This PR also fixes the DefaultOpenTelemetry builder to be it's own class and not a sub-class.
