+++ api/all/src/main/java/io/opentelemetry/api/DefaultOpenTelemetry.java
@@ -5,14 +5,10 @@
 
 package io.opentelemetry.api;
 
-import static java.util.Objects.requireNonNull;
-
 import io.opentelemetry.api.metrics.MeterProvider;
 import io.opentelemetry.api.trace.TracerProvider;
 import io.opentelemetry.context.propagation.ContextPropagators;
 import io.opentelemetry.spi.OpenTelemetryFactory;
-import io.opentelemetry.spi.metrics.MeterProviderFactory;
-import io.opentelemetry.spi.trace.TracerProviderFactory;
 import java.util.ServiceLoader;
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.ThreadSafe;
@@ -45,8 +41,13 @@ public class DefaultOpenTelemetry implements OpenTelemetry {
     globalOpenTelemetry = openTelemetry;
   }
 
-  static Builder builder() {
-    return new Builder();
+  /**
+   * Returns a builder for the {@link DefaultOpenTelemetry}.
+   *
+   * @return a builder for the {@link DefaultOpenTelemetry}.
+   */
+  public static DefaultOpenTelemetryBuilder builder() {
+    return new DefaultOpenTelemetryBuilder();
   }
 
   @Nullable private static volatile OpenTelemetry globalOpenTelemetry;
@@ -93,7 +94,7 @@ public class DefaultOpenTelemetry implements OpenTelemetry {
    * @throws IllegalStateException if a specified provider is not found
    */
   @Nullable
-  private static <T> T loadSpi(Class<T> providerClass) {
+  static <T> T loadSpi(Class<T> providerClass) {
     String specifiedProvider = System.getProperty(providerClass.getName());
     ServiceLoader<T> providers = ServiceLoader.load(providerClass);
     for (T provider : providers) {
@@ -111,58 +112,5 @@ public class DefaultOpenTelemetry implements OpenTelemetry {
   // for testing
   static void reset() {
     globalOpenTelemetry = null;
-  }
-
-  protected static class Builder implements OpenTelemetryBuilder<Builder> {
-    protected ContextPropagators propagators = ContextPropagators.noop();
-
-    protected TracerProvider tracerProvider;
-    protected MeterProvider meterProvider;
-
-    @Override
-    public Builder setTracerProvider(TracerProvider tracerProvider) {
-      requireNonNull(tracerProvider, "tracerProvider");
-      this.tracerProvider = tracerProvider;
-      return this;
-    }
-
-    @Override
-    public Builder setMeterProvider(MeterProvider meterProvider) {
-      requireNonNull(meterProvider, "meterProvider");
-      this.meterProvider = meterProvider;
-      return this;
-    }
-
-    @Override
-    public Builder setPropagators(ContextPropagators propagators) {
-      requireNonNull(propagators, "propagators");
-      this.propagators = propagators;
-      return this;
-    }
-
-    @Override
-    public OpenTelemetry build() {
-      MeterProvider meterProvider = this.meterProvider;
-      if (meterProvider == null) {
-        MeterProviderFactory meterProviderFactory = loadSpi(MeterProviderFactory.class);
-        if (meterProviderFactory != null) {
-          meterProvider = meterProviderFactory.create();
-        } else {
-          meterProvider = MeterProvider.getDefault();
-        }
-      }
-
-      TracerProvider tracerProvider = this.tracerProvider;
-      if (tracerProvider == null) {
-        TracerProviderFactory tracerProviderFactory = loadSpi(TracerProviderFactory.class);
-        if (tracerProviderFactory != null) {
-          tracerProvider = tracerProviderFactory.create();
-        } else {
-          tracerProvider = TracerProvider.getDefault();
-        }
-      }
-
-      return new DefaultOpenTelemetry(tracerProvider, meterProvider, propagators);
-    }
   }
 }
